requirements.txt
ajax==3.0.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
class-registry==2.1.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
crispy==0.7.4
crispy-bootstrap5==2024.2
crispy-templates==0.2.0
cycler==0.12.1
decorator==5.1.1
Django==4.2.13
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-bootstrap5==24.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-extensions==3.2.3
django-filter==24.2
django-rest==0.8.7
django-rest-framework==0.1.0
django-robots==6.1
djangorestframework==3.15.1
dominate==2.9.1
fabio==2024.4.0
filters==1.3.2
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
fonttools==4.53.0
h5py==3.11.0
hdf5plugin==4.4.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
optional-django==0.3.0
packaging==24.1
pillow==10.3.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
pytz==2024.1
react==4.3.0
regex==2024.5.15
requests==2.32.3
silx==2.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
visitor==0.1.3
Werkzeug==3.0.3
zipp==3.19.2
